add utility functions for closing streams
Now that we support half-closed streams, fully closing a stream involves reading
an EOF from the other side. However, doing this manually can be a pain.

Currently, we work around this by:

1. Not waiting for the EOF.
2. "Forgetting" the stream in the swarm on `Close`. This is changing with the
swarm/transport refactor.

However, this is:

1. Incorrect (the stream is still open and could be read from).
2. Could leak the stream if the other side never closes it.
3. Could block the stream if the other side tries to write to it.

So, enter these helper functions. Most of our calls to `Close` will likely be
replaced with `go FullClose(stream)` (or just `FullClose(stream)` if we're
already in a goroutine).